# openapi/openapi.json
{
  "openapi" : "3.1.0",
...
  "components" : {
    "schemas" : {
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "errorId" : {
            "type" : "string",
            "examples" : [ "50905466-1881-457b-b42f-fb7b2bfb1610" ],
            "description" : "The unique identifier that can be used for tracking request and error response."
          },
          "httpStatusCode" : {
            "type" : "integer",
            "format" : "int32",
            "examples" : [ 500 ],
            "description" : "The HTTP status code related to the error message."
          },
          "httpStatusDescription" : {
            "type" : "string",
            "examples" : [ "Internal Server Error" ],
            "description" : "The descriptive name of the HTTP status code."
          },
          "appErrorCode" : {
            "type" : "string",
            "examples" : [ "FDR-0500" ],
            "description" : "The operational error code related to the error response"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorMessage"
            },
            "description" : "The list of specific errors to show for the error response."
          }
        }
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "Metadata" : {
        "type" : "object",
        "properties" : {
          "pageSize" : {
            "type" : "integer",
            "format" : "int32",
            "examples" : [ 25 ],
            "description" : "The size of the current page shown in response"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32",
            "examples" : [ 1 ],
            "description" : "The index of the current page shown in response"
          },
          "totPage" : {
            "type" : "integer",
            "format" : "int32",
            "examples" : [ 3 ],
            "description" : "The total number of the pages that can be retrieved in order to show all elements."
          }
        }
      },
...
      "PaginatedFlowsResponse" : {
        "type" : "object",
        "properties" : {
          "metadata" : {
            "$ref" : "#/components/schemas/Metadata",
            "type" : "object",
            "description" : "The metadata related to the paginated response."
          },
          "count" : {
            "type" : "integer",
            "format" : "int64",
            "examples" : [ 100 ],
            "description" : "The number of elements that can be found in this page."
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FlowByPSP"
            },
            "description" : "The list of flows that are included in this page. Date-time fields inside each item are defined with UTC time-zone."
          }
        }
      },
...
      "SingleFlowResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "$ref" : "#/components/schemas/ReportingFlowStatusEnum",
            "type" : "string",
            "examples" : [ "CREATED" ],
            "description" : "The specific status of publication of the flow"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "examples" : [ 4 ],
            "description" : "The revision (or version) of the flow."
          },
          "created" : {
            "$ref" : "#/components/schemas/Instant",
            "type" : "string",
            "examples" : [ "2025-01-01T12:00:30.900000Z" ],
            "description" : "The date and time on which the flow is created. Defined with UTC time-zone."
          },
          "updated" : {
            "$ref" : "#/components/schemas/Instant",
            "type" : "string",
            "examples" : [ "2025-01-01T12:10:30.900000Z" ],
            "description" : "The last date and time on which the flow is updated. Defined with UTC time-zone."
          },
          "fdr" : {
            "type" : "string",
            "examples" : [ "2025-01-0188888888888-0000001" ],
            "description" : "The unique identifier of the flow."
          },
          "fdrDate" : {
            "$ref" : "#/components/schemas/Instant",
            "type" : "string",
            "examples" : [ "2025-01-01T12:20:30.800000Z" ],
            "description" : "The date related to the flow reporting. Defined with UTC time-zone."
          },
          "regulation" : {
            "type" : "string",
            "examples" : [ "SEPA - Bonifico X" ],
            "description" : "The description related to the regulation payment related to the flow."
          },
          "regulationDate" : {
            "$ref" : "#/components/schemas/LocalDate",
            "type" : "string",
            "examples" : [ "2023-04-03" ],
            "description" : "The date of the regulation payment related to the flow. Business date. The original calendar day is preserved."
          },
          "bicCodePouringBank" : {
            "type" : "string",
            "examples" : [ "UNCRITMMXXX" ],
            "description" : "The BIC code of the bank where the regulation payment related to the flow will be poured."
          },
          "sender" : {
            "$ref" : "#/components/schemas/Sender",
            "type" : "object",
            "description" : "The information related to the entity that will compile and send the published flow."
          },
          "receiver" : {
            "$ref" : "#/components/schemas/Receiver",
            "type" : "object",
            "description" : "The information related to the entity that will receive the published flow."
          },
          "published" : {
            "$ref" : "#/components/schemas/Instant",
            "type" : "string",
            "examples" : [ "2025-01-01T12:20:30.900000Z" ],
            "description" : "The date and time on which the flow is published. Defined with UTC time-zone."
          },
          "computedTotPayments" : {
            "type" : "integer",
            "format" : "int64",
            "examples" : [ 100 ],
            "description" : "The computed total number of payments included in the flow during the compilation."
          },
          "computedSumPayments" : {
            "type" : "number",
            "format" : "double",
            "examples" : [ 100.95 ],
            "pattern" : "^\\d{1,2147483647}[.]\\d{1,2}?$",
            "description" : "The computed total amount of payments calculated in the flow during the compilation."
          },
          "totPayments" : {
            "type" : "integer",
            "format" : "int64",
            "examples" : [ 100 ],
            "description" : "The total number of payments to be included in the flow during the flow compilation."
          },
          "sumPayments" : {
            "type" : "number",
            "format" : "double",
            "examples" : [ 100.95 ],
            "pattern" : "^\\d{1,2147483647}[.]\\d{1,2}?$",
            "description" : "The total amount of payments to be calculated in the flow during the flow compilation."
          }
        }
      }
    }
  },
  "paths" : {
...
    "/organizations/{organizationId}/fdrs" : {
      "get" : {
        "summary" : "Get all published flow related to creditor institution",
        "description" : "## Description:\nThis API permits to search all published flows for a specific creditor institution,\nformatted in a paginated view. The search can be enhanced including the PSP identifier\nin order to filter only the flows for certain PSP. The only flows retrieved are the latest\nrevision, as same as \"nodoChiediElencoFlussiRendicontazione\" primitive does.<br>\nBefore executing the query, the search filters are validated against entities configured for\n<i>Nodo dei Pagamenti</i> environment, in particular on PSP (if that search filter is defined).<br>\nThe result of the query is paginated and contains all the metadata needed for pagination purposes.<br>\n\nQuery date-time filters are defined with UTC time-zone.<br>\n\n\n## Error codes:\nAPPLICATIVE CODE | HTTP CODE | MESSAGE\n- | - | -\n**FDR-2000** | 400 | PSP with ID [%] is invalid or unknown.\n**FDR-2001** | 400 | PSP with ID [%] is not enabled.\n**FDR-2008** | 400 | Creditor institution with ID [%] is invalid or unknown.\n**FDR-2009** | 400 | Creditor institution with ID [%] is not enabled.\n\n\n## API properties:\nPROPERTY | VALUE\n- | -\n***Internal*** | N\n***External*** | Y\n***Synchronous*** | Synchronous\n***Authorization*** | Y (Authorizer)\n***Authentication*** | Y (Subscription Key)\n***TPS*** | 1.0/sec\n***Idempotency*** | Y\n***Stateless*** | Y\n***Read/Write Intensive*** | Read\n***Cacheable*** | Y\n",
        "operationId" : "IOrganizationsController_getAllPublishedFlows",
        "tags" : [ "Organizations" ],
        "parameters" : [ {
          "description" : "The creditor institution identifier, used as a search filter",
          "example" : "15376371009",
          "name" : "organizationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "^(.{1,35})$"
          }
        }, {
          "description" : "A date-time to be used as a lower limit search on flow date. Defined with UTC time-zone and formatted as ISO-8601 (yyyy-MM-dd'T'HH:mm:ss'Z'). If omitted, the server uses a dynamic default equal to the start of the previous 30th day at 00:00:00Z.",
          "example" : "2025-01-01T12:00:00Z",
          "name" : "flowDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "required" : false
        }, {
          "description" : "The index of the page to be shown in the result",
          "example" : 1,
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "default" : 1,
            "minimum" : 1
          }
        }, {
          "description" : "The PSP identifier, used as a search filter",
          "example" : "88888888888",
          "name" : "pspId",
          "in" : "query",
          "schema" : {
            "type" : "string",
            "pattern" : "^(.{1,35})$"
          }
        }, {
          "description" : "A date-time to be used as a lower limit search on publication date. Defined with UTC time-zone and formatted as ISO-8601 (yyyy-MM-dd'T'HH:mm:ss'Z'). If omitted, the server uses a dynamic default equal to the start of the previous 30th day at 00:00:00Z.",
          "example" : "2025-01-01T12:00:00Z",
          "name" : "publishedGt",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "required" : false
        }, {
          "description" : "The number of the elements of the page to be shown in the result",
          "example" : 50,
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "default" : 1000,
            "maximum" : 1000,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginatedFlowsResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "example" : {
                  "httpStatusCode" : 400,
                  "httpStatusDescription" : "Bad Request",
                  "appErrorCode" : "FDR-XXXX",
                  "errors" : [ {
                    "path" : "<detail.path.if-exist>",
                    "message" : "<detail.message>"
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "example" : {
                  "httpStatusCode" : 404,
                  "httpStatusDescription" : "Not Found",
                  "appErrorCode" : "FDR-XXXX",
                  "errors" : [ {
                    "message" : "<detail.message>"
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "example" : {
                  "errorId" : "50905466-1881-457b-b42f-fb7b2bfb1610",
                  "httpStatusCode" : 500,
                  "httpStatusDescription" : "Internal Server Error",
                  "appErrorCode" : "FDR-0500",
                  "errors" : [ {
                    "message" : "An unexpected error has occurred. Please contact support."
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{organizationId}/fdrs/{fdr}/revisions/{revision}/psps/{pspId}" : {
      "get" : {
        "summary" : "Get single published flow related to creditor institution, searching by name and revision",
        "description" : "## Description:\nThis API permits to search a single published flows for a specific creditor institution.\nIn order to do so, it is required to add the following search filters:\n - Creditor institution identifier: for filtering by specific organization\n - PSP identifier: for filtering by flow-related PSP\n - Flow name: for filtering by specific instance of the flow\n - Revision: for filtering by flow revision\n\nBefore executing the query, the search filters are validated against entities configured for\n<i>Nodo dei Pagamenti</i> environment, in particular on creditor institution and PSP. Also,\nthe name of the flow is validated against a specific standard format.<br>\n\nDate-time fields in the response are defined with UTC time-zone. Business-date fields preserve the original calendar day.<br>\n\n\n## Error codes:\nAPPLICATIVE CODE | HTTP CODE | MESSAGE\n- | - | -\n**FDR-2000** | 400 | PSP with ID [%] is invalid or unknown.\n**FDR-2001** | 400 | PSP with ID [%] is not enabled.\n**FDR-2008** | 400 | Creditor institution with ID [%] is invalid or unknown.\n**FDR-2009** | 400 | Creditor institution with ID [%] is not enabled.\n**FDR-3008** | 400 | Flow identifier [%] contains a date that is not compliant.\n**FDR-3009** | 400 | Flow identifier [%] contains a PSP ID that is not compliant.\n**FDR-3001** | 404 | Flow with ID [%] not found.\n\n\n## API properties:\nPROPERTY | VALUE\n- | -\n***Internal*** | N\n***External*** | Y\n***Synchronous*** | Synchronous\n***Authorization*** | Y (Authorizer)\n***Authentication*** | Y (Subscription Key)\n***TPS*** | 1.0/sec\n***Idempotency*** | Y\n***Stateless*** | Y\n***Read/Write Intensive*** | Read\n***Cacheable*** | Y\n",
        "operationId" : "IOrganizationsController_getSinglePublishedFlow",
        "tags" : [ "Organizations" ],
        "parameters" : [ {
          "description" : "The flow name, used as a search filter",
          "example" : "2025-01-0188888888888-0001",
          "name" : "fdr",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The creditor institution identifier, used as a search filter",
          "example" : "15376371009",
          "name" : "organizationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The PSP identifier, used as a search filter",
          "example" : "88888888888",
          "name" : "pspId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "The specific revision of the flow needed",
          "example" : 1,
          "name" : "revision",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleFlowResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "example" : {
                  "httpStatusCode" : 400,
                  "httpStatusDescription" : "Bad Request",
                  "appErrorCode" : "FDR-XXXX",
                  "errors" : [ {
                    "path" : "<detail.path.if-exist>",
                    "message" : "<detail.message>"
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "example" : {
                  "httpStatusCode" : 404,
                  "httpStatusDescription" : "Not Found",
                  "appErrorCode" : "FDR-XXXX",
                  "errors" : [ {
                    "message" : "<detail.message>"
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "example" : {
                  "errorId" : "50905466-1881-457b-b42f-fb7b2bfb1610",
                  "httpStatusCode" : 500,
                  "httpStatusDescription" : "Internal Server Error",
                  "appErrorCode" : "FDR-0500",
                  "errors" : [ {
                    "message" : "An unexpected error has occurred. Please contact support."
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info" : {
    "title" : "FDR - Flussi di Rendicontazione",
...
    "termsOfService" : "https://www.pagopa.gov.it/",
    "version" : "1.2.2"
  },
  "servers" : [ {
    "url" : "https://{host}/",
    "description" : "Environment host",
    "variables" : {
      "host" : {
        "default" : "https://api.dev.platform.pagopa.it/"
      }
    }
  }, {
    "url" : "http://localhost:8080/",
    "description" : "Localhost"
  } ]
}